test: DECK 상수화로 인해 안도는 test 제거
