+++ src/test/java/blackjack/domain/card/CardsTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardNumber;
