+++ src/test/java/blackjack/domain/participant/GameManagerTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import blackjack.domain.game.GameManager;
 import blackjack.domain.participant.Dealer;
@@ -16,6 +16,9 @@ public class GameManagerTest {
         Players players = new Players("pobi,jason", new Dealer());
         GameManager gameManager = new GameManager(players);
         gameManager.giveCards();
+        for(Player player : players.getPlayers()) {
+            assertEquals(player.toList().size(), 2);
+        }
         assertEquals(players.getPlayers().size(), 2);
         assertTrue(players.getPlayers().stream().allMatch(player -> player.toList().size() == 2));
     }
