+++ src/test/java/blackjack/domain/participant/PlayerTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardNumber;
